fix client pool
