+++ autopcr/module/modulemgr.py
@@ -2,12 +2,14 @@ from dataclasses import dataclass
 
 from dataclasses_json import dataclass_json
 from typing import List, Dict
+from abc import abstractmethod
 
 from ..model.error import *
 from ..model.enums import *
 from ..db.database import db
 from .modulebase import Module, ModuleResult, eResultStatus
-
+from autopcr.core.pcrclient import pcrclient
+from 
 import traceback
 
 @dataclass_json
@@ -22,16 +24,68 @@ class TaskResult:
         else:
             return ModuleResult()
 
+@dataclass_json
+@dataclass
+class ResultInfo:
+    alias: str = ""
+    key: str = ""
+    path: str = ""
+    time: str = ""
+    url: str = ""
+    _type: str = ""
+    status: eResultStatus = eResultStatus.SKIP
+
+    def save_result(self, result):
+        with open(self.path, 'w') as f:
+            f.write(result.to_json())
+    def delete_result(self):
+        if os.path.exists(self.path):
+            os.remove(self.path)
+    def get_result(self):
+        raise NotImplementedError
+    def response(self, url_format: str):
+        url = url_format.format(self.alias)
+        return ResultInfo(alias = self.alias, key = self.key, time = self.time, url = url, status = self.status)
+
+@dataclass_json
+@dataclass
+class TaskResultInfo(ResultInfo):
+    _type: str = "daily_result"
+    def get_result(self) -> TaskResult:
+        with open(self.path, 'r') as f:
+            return TaskResult.from_json(f.read())
+
+@dataclass_json
+@dataclass
+class ModuleResultInfo(ResultInfo):
+    _type: str = "single_result"
+    def get_result(self) -> ModuleResult:
+        with open(self.path, 'r') as f:
+            return ModuleResult.from_json(f.read())
+
 class ModuleManager:
     _modules: List[type] = []
 
-    def __init__(self, config, parent):
-        from .modulelistmgr import ModuleListManager
-        from .accountmgr import Account
+    @abstractmethod
+    def get_client() -> pcrclient:
+        pass
 
-        self.parent: Account = parent
+    @abstractmethod
+    async def save_daily_result(self, resp: TaskResult, status: eResultStatus) -> TaskResultInfo:
+        pass
+
+    @abstractmethod
+    async def save_single_result(self, key: str, resp: ModuleResult) -> ModuleResultInfo:
+        pass
+
+    @abstractmethod
+    def is_clan_battle_forbidden(self) -> bool:
+        pass
+
+    def __init__(self, config):
+        from .modulelistmgr import ModuleListManager
         self.modules_list: ModuleListManager = ModuleListManager(self)
-        self.client = self.parent.get_client()
+        self.client = self.get_client()
         self._load_config(config)
     
     def _load_config(self, config):
@@ -66,7 +120,7 @@ class ModuleManager:
             status = eResultStatus.WARNING
         if any(m.status == eResultStatus.PANIC or m.status == eResultStatus.ERROR for m in resp.result.values()):
             status = eResultStatus.ERROR
-        res = await self.parent.save_daily_result(resp, status)
+        res = await self.save_daily_result(resp, status)
         return res
 
     async def do_from_key(self, config: dict, key: str, isAdminCall: bool = False) -> "ModuleResultInfo":
@@ -76,11 +130,11 @@ class ModuleManager:
         modules = [self.modules_list.get_module_from_key(key)]
         raw_resp = await self.do_task(config, modules, isAdminCall)
         resp = raw_resp.result[key] 
-        res = await self.parent.save_single_result(key, resp)
+        res = await self.save_single_result(key, resp)
         return res
 
     async def do_task(self, config: dict, modules: List[Module], isAdminCall: bool = False) -> TaskResult:
-        if db.is_clan_battle_time() and self.parent.is_clan_battle_forbidden() and not isAdminCall:
+        if db.is_clan_battle_time() and self.is_clan_battle_forbidden() and not isAdminCall:
             raise PanicError("会战期间禁止执行任务")
 
         client = self.client
